scion-pki: add tests for cert validate command
Add unit tests for the `scion-pki cert validate` command.

Furthermore, move the the command definition to its own file.

Closes #4176

GitOrigin-RevId: d7b73fed5b66c459bccfad5304b2f0c1a2533df2
