+++ scion-pki/certs/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "observability.go",
         "renew.go",
         "sign.go",
+        "validate.go",
         "verify.go",
     ],
     importpath = "github.com/scionproto/scion/scion-pki/certs",
@@ -60,6 +61,7 @@ go_test(
         "fingerprint_test.go",
         "inspect_test.go",
         "renew_test.go",
+        "validate_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
